+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Progress Bar to Attachment uploads
+
 ## [2.2.0] - 2022-09-27
 
 ### Fixed