+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Progress Bar to Attachment uploads
+- `continueWithUploadingAttachments` prop to `<OneBlinkForm />`
 
 ## [2.2.0] - 2022-09-27