+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Progress Bar to Attachment uploads
-- `continueWithUploadingAttachments` prop to `<OneBlinkForm />`
+- Add prompt to allow continuing submission whilst attachments are uploading
 
 ## [2.2.0] - 2022-09-27