+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Signature element button "done" to "save signature" and validation message
+
 ## [2.3.0] - 2022-10-12
 
 ### Added